requirements.txt
fastapi~=0.90.1
openai~=0.26.5
pyrebase4